Project MoneyBall
Moneyball Book
The central premise of book Moneyball is that the collective wisdom of
baseball insiders (including players, managers, coaches, scouts, and the
front office) over the past century is subjective and often flawed.
Statistics such as stolen bases, runs batted in, and batting average,
typically used to gauge players, are relics of a 19th-century view of
the game and the statistics available at that time. The book argues that
the Oakland A's' front office took advantage of more analytical gauges
of player performance to field a team that could better compete against
richer competitors in Major League Baseball (MLB).
Rigorous statistical analysis had demonstrated that on-base percentage
and slugging percentage are better indicators of offensive success, and
the A's became convinced that these qualities were cheaper to obtain on
the open market than more historically valued qualities such as speed
and contact. These observations often flew in the face of conventional
baseball wisdom and the beliefs of many baseball scouts and executives.
By re-evaluating the strategies that produce wins on the field, the 2002
Athletics, with approximately US 44 million dollars in salary, were
competitive with larger market teams such as the New York Yankees, who
spent over US$125 million in payroll that same season.