Simple Linear Regression Project
For this project I used Bike Sharing Demand Kaggle challenge!
Exploratory Data Analysis and begin to get an understanding that
sometimes certain models are not a good choice for a data set. In this
case, we will discover that Linear Regression may not be the best choice
given our data